JWT access token 1 day expiry token latest update, change recipeDataId datatype from int to Long

+++ recipeapp_backend/src/main/java/org/forksmash/recipeapp_backend/recipe/Recipe.java
@@ -47,7 +47,7 @@ public class Recipe {
     private UserProfile userProfile; 
 
     @Column(name = "recipe_data_id")
-    private int recipeDataId;
+    private Long recipeDataId;
 
     @NotNull(message = "An ingredient type must be assigned to an ingredient")
     //@Convert(converter = JpaConverterJson.class)
@@ -64,7 +64,7 @@ public class Recipe {
     @Column(name = "nutrition", columnDefinition = "json")
     private String nutrition;
 
-    public Recipe(int recipeDataId, String info, String instructions, String extendedIngredients, String nutrition, UserProfile userProfile) {
+    public Recipe(Long recipeDataId, String info, String instructions, String extendedIngredients, String nutrition, UserProfile userProfile) {
         this.recipeDataId = recipeDataId;
         this.info = info;
         this.instructions = instructions;
